requirements.txt
absl-py==2.1.0
aiohttp==3.9.1
aiosignal==1.3.1
annotated-types==0.6.0
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astor==0.8.1
asttokens==2.4.1
async-lru==2.0.4
attrs==23.2.0
Babel==2.14.0
backoff==2.2.1
beautifulsoup4==4.12.3
bleach==6.1.0
blessed==1.20.0
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.2
ci-info==0.3.0
click==8.1.7
comm==0.2.1
configobj==5.0.9
configparser==7.1.0
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
distlib==0.3.7
distro==1.9.0
editor==1.6.5
etelemetry==0.3.1
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.12.4
fitz==0.0.1.dev2
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
fsspec==2023.12.2
git-python==1.0.3
gitdb==4.0.11
GitPython==3.1.41
h11==0.14.0
h5py==3.11.0
html2image==2.0.4.3
httpcore==1.0.2
httplib2==0.22.0
httpx==0.26.0
huggingface-hub==0.20.2
idna==3.6
imageio==2.36.0
importlib-metadata==7.0.1
importlib_resources==6.4.5
inquirer==3.2.1
ipykernel==6.28.0
ipython==8.20.0
ipywidgets==8.1.1
isodate==0.6.1
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.9.0
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.12.5
jupyter_server_terminals==0.5.2
jupyterlab==4.1.0
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.2
keras==3.3.3
kiwisolver==1.4.5
lazy_loader==0.4
litellm==1.17.9
looseversion==1.3.0
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.3
matplotlib==3.8.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
ml-dtypes==0.4.0
monotonic==1.6
multidict==6.0.4
namex==0.0.8
nbclient==0.9.0
nbconvert==7.15.0
nbextensions==66.0.2
nbformat==5.9.2
nest-asyncio==1.5.9
networkx==3.4.2
nibabel==5.3.2
nipype==1.8.6
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
openai==1.7.2
optree==0.11.0
overrides==7.7.0
packaging==23.2
pandas==2.2.3
pandocfilters==1.5.0
parso==0.8.3
pathlib==1.0.1
pexpect==4.9.0
pillow==10.2.0
pipenv==2023.10.3
platformdirs==3.11.0
posthog==3.3.1
prometheus-client==0.19.0
prompt-toolkit==3.0.43
protobuf==4.24.3
prov==2.0.1
psutil==5.9.7
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
pydantic==2.5.3
pydantic_core==2.14.6
pydot==3.0.2
Pygments==2.17.2
PyMuPDF==1.24.12
pyparsing==3.1.1
pystache==0.6.5
python-dateutil==2.8.2
python-dotenv==1.0.0
python-json-logger==2.0.7
pytils==0.4.1
pytz==2024.2
pyxnat==1.6.2
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
rdflib==6.3.2
readchar==4.0.5
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.7.0
rpds-py==0.17.1
runs==1.2.0
scikit-image==0.24.0
scipy==1.14.1
seaborn==0.13.2
Send2Trash==1.8.2
simplejson==3.19.3
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
static==1.1.1
terminado==0.18.0
tifffile==2024.9.20
tiktoken==0.4.0
tinycss2==1.2.1
tokenizers==0.15.0
tokentrim==0.1.13
toml==0.10.2
tools==0.1.9
tornado==6.4
tqdm==4.66.1
traitlets==5.14.1
traits==6.3.2
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.1.0
virtualenv==20.24.5
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
wget==3.2
widgetsnbextension==4.0.9
xmod==1.8.1
yarl==1.9.4
zipp==3.17.0
